requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==43.0.0
decorator==5.1.1
Deprecated==1.2.14
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.0.7
django-allauth==64.0.0
django-allauth-ui==1.2.0
django-cache-url==3.4.5
django-cleanup==8.1.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-widget-tweaks==1.5.0
environs==11.0.0
exceptiongroup==1.2.2
fabric==3.2.2
idna==3.7
iniconfig==2.0.0
invoke==2.2.0
MarkupSafe==2.1.5
marshmallow==3.21.3
oauthlib==3.2.2
packaging==24.1
paramiko==3.4.0
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
pytest==8.3.1
pytest-django==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
slippers==0.6.2
sqlparse==0.5.1
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
